canny_edge_detection_user.c
The variation of the Canny edge-detection used in this exercise will be comprised of five stages. The first stage is an RGB-to-grayscale transformation to simplify the number of bits per pixel to simplify the image processing. The second stage is applying a Gaussian smoothing to reduce the noise which would perpetuate to the result. The main edge detection occurs in stage 3, where the partial derivatives of each pixel is calculated to conclude to the gradient intensity. The edges are then refined with a Non-Maximum suppression in stage 4. Finally pixels that do not belong to any edges or weak are removed through a hysteresis algorithm. 

The script uses a video driver kernel to display on a VGA screen. This part can be ommitted.